eAttest updates: V2, V3 and various fixes

+++ src/main/kotlin/org/taktik/freehealth/middleware/service/EattestV3Service.kt
@@ -71,7 +71,7 @@ interface EattestV3Service {
         patientFirstName: String,
         patientLastName: String,
         patientGender: String,
-        referenceDate: Int?,
+        referenceDate: Long?,
         eAttestRef: String,
         reason: String,
         attemptNbr: Int?
